Add PushLayer and OnEvent methods to Application class
Added the `PushLayer` method to the `Motion::App::Application`
class, which accepts a `std::shared_ptr` to a `Motion::Core::Layer`.
The method is currently unimplemented.

Added the `OnEvent` method to the same class, which processes
events using an `EventHandler`. It dispatches `EventWindowClose`
and `EventWindowResize<uint32_t>` events using the `EVENT_CALLBACK`
macro, binding them to `OnWindowClose` and `OnWindowResize`
callback functions.